Merged in feature/PE-219_pointwise_conv_exhaustive_tests (pull request #68)
PE-219: Introducing pytest + pointwise convolution exhaustive test

Introducing Pytest with a test for pointwise convolution configurations used in MobileNetV2.

I refactored the TestBase class that is the base for the new unit tests so that it can run with either int32 or float32 as the target type, I also moved the assert functions to the global scope.

Approved-by: Maxim Karpushin